moved test resources into test subfolder
