+++ package.json
@@ -9,6 +9,7 @@
     "build": "tsc -p tsconfig.json",
     "setup": "bin/setup.sh",
     "start": "node --require dotenv/config dist/index.js",
+    "postinstall": "rm node_modules/mongodb/mongodb.d.ts",
     "postbuild": "npm run copy-files",
     "copy-files": "copyfiles -u 1 src/**/*.yaml src/templates/* dist/"
   },
