+++ src/test/app.spec.ts
@@ -8,14 +8,14 @@ import LRU from 'lru-cache';
 import { Collection } from 'mongodb';
 import { MongoMemoryServer } from 'mongodb-memory-server';
 import { Issuer } from 'openid-client';
-import { build } from './app';
-import { resetConfig } from './config';
-import * as IssuerHelper from './issuer';
-import * as Certificate from './templates/Certificate';
-import demoCredential from './demoCredential.json';
-import { AuthType } from './issuer';
-import * as Database from './database';
-import { dbCreate, dbConnect, dbDisconnect, Credential } from '../helpers/database';
+import { build } from '../app';
+import { resetConfig } from '../config';
+import * as IssuerHelper from '../issuer';
+import * as Certificate from '../templates/Certificate';
+import demoCredential from '../demoCredential.json';
+import { AuthType } from '../issuer';
+import * as Database from '../database';
+import { dbCreate, dbConnect, dbDisconnect, Credential } from './database';
 
 const sandbox = createSandbox();
 const lruStub = sandbox.createStubInstance(LRU) as SinonStubbedInstance<LRU> & LRU;
